Se adapto los permisos en el UpdDel de los controladores de alistamientos personas personaDocumentos personaLicencias personasRh vehiculos y vehiculosDocumentos y de igual forma en el list de cada una de los ya mencionados

+++ app/Controllers/AlistamientosController.php
@@ -409,6 +409,9 @@ class AlistamientosController extends BaseController{
 		$roles = null; $tiposdocumentos=null; $personas=null; $numeroDePaginas=null; $id=null; $boton=null;
 		$quiereActualizar = false; $ruta='alistamientosList.twig'; $responseMessage = null;
 		$generos=null; $estadocivil=null; $rh=null; $niveleducativo=null;
+		$mensajeNoPermisos='Su rol no tiene permisos para realizar esta funcion';
+
+		$sessionUserPermission = $_SESSION['userLicense'] ?? null;
 
 		if($request->getMethod()=='POST'){
 			$postData = $request->getParsedBody();
@@ -420,12 +423,19 @@ class AlistamientosController extends BaseController{
 				$id = $postData['id'] ?? null;				
 				if ($id) {
 					if ($btnDocumentos == 'doc') {
+					  if (in_array('documentslist', $sessionUserPermission)) {
 						$DocumentosController = new PersonaDocumentosController();
 						return $DocumentosController->listPersonasDocumentos($id);
-
+					  }else{
+						$responseMessage=$mensajeNoPermisos;
+					  }
 					}elseif ($btnDocumentos == 'lic') {
+					  if (in_array('licenselist', $sessionUserPermission)) {
 						$LicenciasController = new PersonaLicenciasController();
 						return $LicenciasController->listPersonasLicencias($id);
+					  }else{
+					  	$responseMessage=$mensajeNoPermisos;
+					  }
 					}else{
 						$responseMessage = 'Opcion no definida, btn: '.$btnDocumentos;
 					}	
@@ -441,6 +451,7 @@ class AlistamientosController extends BaseController{
 			}
 			if ($id) {
 				if($boton == 'del'){
+				 if (in_array('checkdel', $sessionUserPermission)) {
 				  try{
 					$people = new Personas();
 					$people->destroy($id);
@@ -454,8 +465,15 @@ class AlistamientosController extends BaseController{
 						$responseMessage= 'Error, No se puede eliminar, '.$prevMessage;
 					}
 				  }
+				 }else{
+				 	$responseMessage=$mensajeNoPermisos;
+				 }
 				}elseif ($boton == 'upd') {
+				  if (in_array('checkupdate', $sessionUserPermission)) {
 					$quiereActualizar=true;
+				  }else{
+				  	$responseMessage=$mensajeNoPermisos;
+				  }
 				}
 			}else{
 				$responseMessage = 'Debe Seleccionar una persona';
