+++ app/Controllers/PersonaDocumentosController.php
@@ -396,6 +396,9 @@ class PersonaDocumentosController extends BaseController{
 	public function postUpdDelDocumentos($request){
 		$tiposdocumentos=null; $documentos=null; $responseMessage = null; $id=null; $boton=null;
 		$quiereActualizar = false; $ruta='personaDocumentosList.twig'; $numeroDePaginas=null; $persona=null;
+		$mensajeNoPermisos='Su rol no tiene permisos para realizar esta funcion';
+
+		$sessionUserPermission = $_SESSION['userLicense'] ?? null;
 
 		if($request->getMethod()=='POST'){
 			$postData = $request->getParsedBody();
@@ -409,6 +412,7 @@ class PersonaDocumentosController extends BaseController{
 			}
 			if ($id) {
 				if($boton == 'del'){
+				 if (in_array('peopledel', $sessionUserPermission)) {
 				  try{
 					$people = new PersonaDocumentos();
 					$people->destroy($id);
@@ -422,8 +426,15 @@ class PersonaDocumentosController extends BaseController{
 						$responseMessage= 'Error, No se puede eliminar, '.$prevMessage;
 					}
 				  }
+				 }else{
+				 	$responseMessage=$mensajeNoPermisos;
+				 }
 				}elseif ($boton == 'upd') {
+				  if (in_array('peopleupd', $sessionUserPermission)) {
 					$quiereActualizar=true;
+				  }else{
+				  	$responseMessage=$mensajeNoPermisos;
+				  }
 				}
 			}else{
 				$responseMessage = 'Debe Seleccionar un documento';
